requirements.txt
streamlit>=1.32.0
google-analytics-data>=0.18.0
google-auth>=2.27.0
pandas>=2.1.0
plotly>=5.18.0
anthropic>=0.25.0
